Add files via upload
This tiny script will offboard Windows 10 machines (in a machines list) from the MDATP portal.  Note - if the Windows Defender Advanced Threat Protection service is restarted on the machine(s), the machine will re-onboard.  This is by design. 

You can set the sense service to disabled in the registry to prevent the service from restarting.